refactor: MVC 패턴 참고하여 패키지 구조 변경
