+++ src/main/java/src/domain/Game.java
@@ -1,4 +1,4 @@
-package src;
+package src.domain;
 
 import java.util.HashSet;
 import java.util.Set;
