+++ src/main/java/src/domain/Lotto.java
@@ -1,4 +1,4 @@
-package src;
+package src.domain;
 
 import java.util.ArrayList;
 import java.util.List;
