+++ src/main/java/src/domain/Money.java
@@ -1,4 +1,4 @@
-package src;
+package src.domain;
 
 public class Money {
