+++ src/main/java/src/domain/WinningAmount.java
@@ -1,4 +1,4 @@
-package src;
+package src.domain;
 
 import java.util.Arrays;
