+++ src/test/java/src/domain/GamesTest.java
@@ -1,6 +1,9 @@
-package src;
+package src.domain;
 
 import org.junit.jupiter.api.Test;
+import src.domain.Game;
+import src.domain.GameNumber;
+import src.domain.Games;
 
 import java.util.List;
 import java.util.Set;
